refactor banners
