+++ src/components/app-banners.tsx
@@ -2,10 +2,9 @@
 
 import Image from "next/image";
 import dynamic from "next/dynamic";
-import { HTMLAttributes, useEffect, useRef, useState } from "react";
-import { twMerge } from "tailwind-merge";
+import { HTMLAttributes, useState } from "react";
 import { Media } from "@/payload-types";
-import { HSCarousel } from "preline/preline";
+import AppCarousel from "@/components/app-carousel";
 
 const AppBannerVideo = dynamic(() => import("@/components/app-banner-video"), {
   ssr: false,
@@ -26,130 +25,50 @@ interface AppBannersProps extends HTMLAttributes<HTMLElement> {
 export default function AppBanners(props: AppBannersProps) {
   const { banners, className } = props;
 
-  const [currentIndex, setCurrentIndex] = useState(0);
-  const timeoutRef = useRef<NodeJS.Timeout | null>(null);
-  const [durations, setDurations] = useState<Record<number, number>>(
-    Object.fromEntries(banners.map((_, index) => [index, PHOTO_DURATION])),
+  const [durations, setDurations] = useState<number[]>(
+    banners.map(() => PHOTO_DURATION),
   );
 
   const handleLoadedMetadata =
     (index: number) => (event: React.SyntheticEvent<HTMLVideoElement>) => {
-      setDurations({
-        ...durations,
-        [index]: event.currentTarget.duration * 1000,
-      });
+      setDurations([
+        ...durations.slice(0, index),
+        event.currentTarget.duration * 1000,
+        ...durations.slice(index),
+      ]);
     };
 
-  function resetTimeout() {
-    if (timeoutRef.current) {
-      clearTimeout(timeoutRef.current);
-    }
-  }
-
-  useEffect(() => {
-    resetTimeout();
-    timeoutRef.current = setTimeout(() => {
-      const newIndex =
-        currentIndex === banners.length - 1 ? 0 : currentIndex + 1;
-      setCurrentIndex(newIndex);
-      const carousel = HSCarousel.getInstance(`#${CAROUSEL_ID}`) as HSCarousel;
-      carousel.goTo(newIndex);
-    }, durations[currentIndex] ?? PHOTO_DURATION);
-
-    return () => resetTimeout();
-  }, [banners.length, currentIndex, durations]);
-
   return (
-    <div
-      id={CAROUSEL_ID}
-      data-hs-carousel={JSON.stringify({
-        loadingClasses: "opacity-0",
-        dotsItemClasses:
-          "hs-carousel-active:bg-primary-700 hs-carousel-active:border-primary-700 size-3 border border-gray-400 rounded-full cursor-pointer dark:border-neutral-600 dark:hs-carousel-active:bg-primary-500 dark:hs-carousel-active:border-primary-500",
-        isInfiniteLoop: true,
-      })}
-      className={twMerge("relative aspect-[4/1.2] w-full", className)}
-    >
-      <div className="hs-carousel relative h-full w-full overflow-hidden bg-white">
-        <div className="hs-carousel-body absolute start-0 top-0 bottom-0 flex flex-nowrap opacity-0 transition-transform duration-700">
-          {banners
-            .filter((banner) => typeof banner.url === "string")
-            .map((banner, index) => (
-              <div key={index} className="hs-carousel-slide">
-                {checkVideo(banner.url!) ? (
-                  <AppBannerVideo
-                    id={`banner-video-${index}`}
-                    className="h-full w-full object-cover"
-                    src={banner.url!}
-                    autoPlay
-                    crossOrigin="anonymous"
-                    loop
-                    muted
-                    playsInline
-                    preload="metadata"
-                    onLoadedMetadata={handleLoadedMetadata(index)}
-                  />
-                ) : (
-                  <Image
-                    className="h-full w-full object-cover"
-                    src={banner.url!}
-                    alt={banner.alt || ""}
-                    sizes="100%"
-                    width={0}
-                    height={0}
-                  />
-                )}
-              </div>
-            ))}
-        </div>
-      </div>
-
-      <button
-        type="button"
-        className="hs-carousel-prev hs-carousel-disabled:opacity-50 absolute inset-y-0 start-0 inline-flex h-full w-11.5 items-center justify-center rounded-s-lg text-gray-800 hover:bg-gray-800/10 focus:bg-gray-800/10 focus:outline-hidden disabled:pointer-events-none dark:text-white dark:hover:bg-white/10 dark:focus:bg-white/10"
-      >
-        <span className="text-2xl" aria-hidden="true">
-          <svg
-            className="size-5 shrink-0"
-            xmlns="http://www.w3.org/2000/svg"
-            width="24"
-            height="24"
-            viewBox="0 0 24 24"
-            fill="none"
-            stroke="currentColor"
-            strokeWidth="2"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          >
-            <path d="m15 18-6-6 6-6"></path>
-          </svg>
-        </span>
-        <span className="sr-only">Previous</span>
-      </button>
-      <button
-        type="button"
-        className="hs-carousel-next hs-carousel-disabled:opacity-50 absolute inset-y-0 end-0 inline-flex h-full w-11.5 items-center justify-center rounded-e-lg text-gray-800 hover:bg-gray-800/10 focus:bg-gray-800/10 focus:outline-hidden disabled:pointer-events-none dark:text-white dark:hover:bg-white/10 dark:focus:bg-white/10"
-      >
-        <span className="sr-only">Next</span>
-        <span className="text-2xl" aria-hidden="true">
-          <svg
-            className="size-5 shrink-0"
-            xmlns="http://www.w3.org/2000/svg"
-            width="24"
-            height="24"
-            viewBox="0 0 24 24"
-            fill="none"
-            stroke="currentColor"
-            strokeWidth="2"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          >
-            <path d="m9 18 6-6-6-6"></path>
-          </svg>
-        </span>
-      </button>
-
-      <div className="hs-carousel-pagination absolute start-0 end-0 bottom-3 flex justify-center space-x-2"></div>
-    </div>
+    <AppCarousel id={CAROUSEL_ID} durations={durations} className={className}>
+      {banners
+        .filter((banner) => typeof banner.url === "string")
+        .map((banner, index) => (
+          <div key={index} className="hs-carousel-slide">
+            {checkVideo(banner.url!) ? (
+              <AppBannerVideo
+                id={`banner-video-${index}`}
+                className="h-full w-full object-cover"
+                src={banner.url!}
+                autoPlay
+                crossOrigin="anonymous"
+                loop
+                muted
+                playsInline
+                preload="metadata"
+                onLoadedMetadata={handleLoadedMetadata(index)}
+              />
+            ) : (
+              <Image
+                className="h-full w-full object-cover"
+                src={banner.url!}
+                alt={banner.alt || ""}
+                sizes="100%"
+                width={0}
+                height={0}
+              />
+            )}
+          </div>
+        ))}
+    </AppCarousel>
   );
 }
